Add helper to roll item quality
