+++ src/features/loot/qualityWeights.js
@@ -1,0 +1,10 @@
+export function rollQualityKey(weights = { normal: 80, magic: 15, rare: 5 }, rng = Math.random) {
+  const entries = Object.entries(weights);
+  const total = entries.reduce((sum, [, w]) => sum + w, 0);
+  let r = rng() * total;
+  for (const [key, weight] of entries) {
+    r -= weight;
+    if (r <= 0) return key;
+  }
+  return entries[0][0];
+}
